+++ HISTORY.md
@@ -9,6 +9,7 @@
 
 ### 2021-06-15
 * threads: added exception handling primitives for try/catch
+* extensions: added try-catch extension primitives
 
 ### 2021-06-15
 * extensions: tweaked world-map primitives